+++ here-naksha-lib-core/src/main/java/com/here/naksha/lib/core/NakshaVersion.java
@@ -76,12 +76,13 @@ public class NakshaVersion implements Comparable<NakshaVersion> {
   public static final String v2_2_16 = "2.2.16";
   public static final String v2_2_17 = "2.2.17";
   public static final String v2_2_18 = "2.2.18";
+  public static final String v2_2_19 = "2.2.19";
 
   /**
    * The latest version of the naksha-extension stored in the resources.
    */
   @AvailableSince(v2_0_5)
-  public static final NakshaVersion latest = of(v2_2_17);
+  public static final NakshaVersion latest = of(v2_2_19);
 
   private final int major;
   private final int minor;