+++ here-naksha-lib-mm-util/src/main/java/com/here/naksha/mom10/DeltaProperties.java
@@ -19,6 +19,7 @@
 package com.here.naksha.mom10;
 
 class DeltaProperties {
+
   private DeltaProperties() {}
 
   /*